fix: add support for gpt-5
Adds support for gpt-5 to the openai provider. Along with this change
i've also added support for `max_tokens` diagnostics. This will show a
warning to the user when max_length is reached. gpt-5 uses thinking inferred
from the max_tokens so we often hit max_length with a non trivial max_tokens
set.